// frontend/src/app/data-management/file-management/shared/enum/access-mask.enum.ts
export enum AccessMaskEnum {
  READ_DATA = '0x1',
  WRITE_DATA = '0x2',
  APPEND_DATA = '0x4',
  READ_EA = '0x8',
  WRITE_EA = '0x10',
  EXECUTE_TRAVERSE = '0x20',
  DELETE_CHILD = '0x40',
  READ_ATTRIBUTES = '0x80',
  WRITE_ATTRIBUTES = '0x100',
  DELETE = '0x10000',
  READ_CONTROL = '0x20000',
  WRITE_AC = '0x40000',
  WRITE_OWNER = '0x80000',
  SYNCHRONIZE = '0x100000',
  ACCESS_SYS_SEC = '0x1000000'
}